New fonctions

+++ src/Controller/TaskController.php
@@ -7,6 +7,7 @@ use App\Form\TaskType;
 use App\Repository\TaskRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -14,26 +15,32 @@ use Symfony\Component\HttpFoundation\Response;
 class TaskController extends AbstractController
 {
     #[Route(path: '/tasks', name: 'task_list')]
+    #[IsGranted("ROLE_USER", message: 'Page not found.')]
     public function listAction(TaskRepository $taskRepository): Response
     {
         return $this->render('task/list.html.twig', ['tasks' => $taskRepository->findAll()]);
     }
 
     #[Route(path: '/tasks/create', name: 'task_create')]
+    #[IsGranted("ROLE_USER", message: 'Page not found.')]
     public function createAction(Request $request, EntityManagerInterface $em): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('login');
+        }
+
         $task = new Task();
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-
+            // add task owner to the database
+            $task->setUser($this->getUser());
             $em->persist($task);
             $em->flush();
 
             $this->addFlash('success', 'La tâche a été bien été ajoutée.');
-
             return $this->redirectToRoute('task_list');
         }
 
@@ -41,10 +48,15 @@ class TaskController extends AbstractController
     }
 
     #[Route(path: '/tasks/{id}/edit', name: 'task_edit')]
+    #[IsGranted("ROLE_USER", message: 'Page not found.')]
     public function editAction(Task $task, Request $request, EntityManagerInterface $em): Response
     {
-        $form = $this->createForm(TaskType::class, $task);
+        //check logged user is same as task's owner
+        if ($this->getUser() !== $task->getUser()) {
+            return $this->redirectToRoute('task_list');
+        }
 
+        $form = $this->createForm(TaskType::class, $task);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -62,8 +74,14 @@ class TaskController extends AbstractController
     }
 
     #[Route(path: '/tasks/{id}/toggle', name: 'task_toggle')]
+    #[IsGranted("ROLE_USER", message: 'Page not found')]
     public function toggleTaskAction(Task $task, EntityManagerInterface $em): Response
     {
+        //check logged user is same as task's owner
+        if ($this->getUser() !== $task->getUser()) {
+            $this->addFlash('danger', "Vous n'avez pas le droit de modifier la tâche.");
+            return $this->redirectToRoute('task_list');
+        }
         $task->toggle(!$task->isDone());
         $em->flush();
 
@@ -73,8 +91,24 @@ class TaskController extends AbstractController
     }
 
     #[Route(path: '/tasks/{id}/delete', name: 'task_delete')]
-    public function deleteTaskAction(Task $task, EntityManagerInterface $em): Response
+    public function deleteTaskAction(int $id,  EntityManagerInterface $em, TaskRepository $taskRepository): Response
     {
+        $task = $taskRepository->find($id);
+        if (empty($task)) {
+            $this->addFlash('danger', 'Page not found');
+            return $this->redirectToRoute('task_list');
+        }
+        if ($task->getUser()->getUsername() === 'anonyme' && !in_array('ROLE_ADMIN', $this->getUser()->getRoles())) {
+            $this->addFlash('danger', 'Only admin can delete this task');
+            return $this->redirectToRoute('task_list');
+        }
+
+        //check logged user is same as task's owner
+        if ($this->getUser() !== $task->getUser() && !$task->getUser()->getUsername() === 'anonyme') {
+            $this->addFlash('danger', "Vous n'avez pas le droit de supprimer la tâche.");
+            return $this->redirectToRoute('task_list');
+        }
+
         $em->remove($task);
         $em->flush();
